mcp_servers/powerpoint/mcp_servers/slides_server/tools/create_slides.py
import os
from collections.abc import Iterable
from typing import Any

from models.response import CreateDeckResponse
from models.slide import PresentationMetadata, SlideDefinition
from models.tool_inputs import CreateDeckInput
from pptx import Presentation
from pptx.presentation import Presentation as PresentationObject
from pydantic import TypeAdapter
from utils.decorators import make_async_background

SLIDES_ROOT = os.getenv("APP_SLIDES_ROOT") or os.getenv("APP_FS_ROOT", "/filesystem")

SLIDE_ADAPTER = TypeAdapter(SlideDefinition)

_LAYOUT_MAP = {
    "title": 0,
    "title_and_content": 1,
    "section_header": 2,
    "two_content": 3,
    "title_only": 5,
    "blank": 6,
}


def _resolve_under_root(path: str) -> str:
    """Map path to the slides root."""
    path = path.lstrip("/")
    full_path = os.path.join(SLIDES_ROOT, path)
    return os.path.normpath(full_path)


def _normalize_directory(directory: str) -> str:
    return "" if directory == "/" else directory.rstrip("/")


def _apply_metadata(
    presentation: PresentationObject, metadata: PresentationMetadata
) -> None:
    core = presentation.core_properties
    if metadata.title is not None:
        core.title = metadata.title
    if metadata.subject is not None:
        core.subject = metadata.subject
    if metadata.author is not None:
        core.author = metadata.author
    if metadata.comments is not None:
        core.comments = metadata.comments


def _populate_title(slide: Any, title: str | None) -> None:
    if not title:
        return
    title_shape = getattr(slide.shapes, "title", None)
    if title_shape and getattr(title_shape, "text_frame", None):
        title_shape.text = title


def _populate_subtitle(slide: Any, subtitle: str | None) -> None:
    if not subtitle:
        return
    body_placeholder = _get_placeholder(slide, "body")
    text_frame = _get_text_frame(body_placeholder)
    if text_frame is not None:
        text_frame.text = subtitle


def _get_text_frame(shape: Any):
    if getattr(shape, "has_text_frame", False):
        return shape.text_frame
    return None


def _populate_bullets(text_frame: Any, bullets: Iterable[str]) -> None:
    if text_frame is None:
        return
    text_frame.clear()
    for idx, item in enumerate(bullets):
        paragraph = text_frame.add_paragraph() if idx > 0 else text_frame.paragraphs[0]
        paragraph.text = item
        paragraph.level = 0


def _populate_table(
    slide: Any, placeholder: Any, rows: list[list[str]], header: bool
) -> None:
    from pptx.util import Inches

    if placeholder is not None:
        left = placeholder.left
        top = placeholder.top
        width = placeholder.width
        height = placeholder.height
    else:
        left = Inches(0.5)
        top = Inches(1.5)
        width = Inches(9)
        height = Inches(5)

    graphic_frame = slide.shapes.add_table(
        len(rows), len(rows[0]), left, top, width, height
    )
    table = graphic_frame.table
    for r, row_values in enumerate(rows):
        for c, cell_value in enumerate(row_values):
            table.cell(r, c).text = cell_value
    if header:
        for cell in table.rows[0].cells:
            for paragraph in cell.text_frame.paragraphs:
                for run in paragraph.runs:
                    run.font.bold = True


def _populate_two_content(slide, columns) -> None:
    left_placeholder = _get_text_frame(_get_placeholder(slide, "left"))
    right_placeholder = _get_text_frame(_get_placeholder(slide, "right"))
    if columns.left and left_placeholder is not None:
        _populate_bullets(left_placeholder, columns.left.items)
    if columns.right and right_placeholder is not None:
        _populate_bullets(right_placeholder, columns.right.items)


def _get_placeholder(slide, placeholder_key: str):
    placeholder_map = {"title": 0, "body": 1, "left": 1, "right": 2}
    idx = placeholder_map.get(placeholder_key)
    if idx is None:
        return None
    try:
        return slide.shapes.placeholders[idx]
    except (IndexError, KeyError):
        return None


def _populate_notes(slide, notes: str | None) -> None:
    if not notes:
        return
    notes_frame = slide.notes_slide.notes_text_frame  # notes_slide always exists
    notes_frame.clear()
    notes_frame.text = notes


@make_async_background
def create_deck(request: CreateDeckInput) -> CreateDeckResponse:
    """Create a new PowerPoint presentation with specified slides and layouts.

    Notes:
        - Directory created if doesn't exist
        - Minimum 1 slide required
        - All template slides removed; only specified slides included

        Layout-specific:
        - Bullets: 'title_and_content', 'two_content' only
        - Tables: header=True makes first row bold
        - two_content: 'columns' overrides 'bullets'/'table' if both provided
    """

    def error(msg: str) -> CreateDeckResponse:
        return CreateDeckResponse(success=False, error=msg)

    try:
        slide_models = [SLIDE_ADAPTER.validate_python(item) for item in request.slides]
    except Exception as exc:
        return error(f"Invalid slides payload: {exc}")

    presentation = Presentation()

    if request.metadata:
        try:
            metadata_model = PresentationMetadata.model_validate(request.metadata)
        except Exception as exc:
            return error(f"Invalid metadata: {exc}")
        _apply_metadata(presentation, metadata_model)

    while presentation.slides:
        presentation.slides._sldIdLst.remove(presentation.slides._sldIdLst[0])

    for slide_model in slide_models:
        layout_index = _LAYOUT_MAP[slide_model.layout]
        slide_layout = presentation.slide_layouts[layout_index]
        slide = presentation.slides.add_slide(slide_layout)

        _populate_title(slide, slide_model.title)
        if slide_model.layout in {"title", "section_header"}:
            _populate_subtitle(slide, slide_model.subtitle)

        if slide_model.layout == "title_and_content":
            body_placeholder = _get_placeholder(slide, "body")
            text_frame = _get_text_frame(body_placeholder)
            if slide_model.bullets and text_frame is not None:
                _populate_bullets(text_frame, slide_model.bullets.items)
            elif slide_model.table:
                _populate_table(
                    slide,
                    body_placeholder,
                    slide_model.table.rows,
                    slide_model.table.header,
                )

        elif slide_model.layout == "two_content":
            if slide_model.columns:
                _populate_two_content(slide, slide_model.columns)
            elif slide_model.bullets:
                left_text_frame = _get_text_frame(_get_placeholder(slide, "left"))
                if left_text_frame is not None:
                    _populate_bullets(left_text_frame, slide_model.bullets.items)
            if slide_model.table:
                right_placeholder = _get_placeholder(slide, "right")
                _populate_table(
                    slide,
                    right_placeholder,
                    slide_model.table.rows,
                    slide_model.table.header,
                )

        if slide_model.notes:
            _populate_notes(slide, slide_model.notes)

    file_path = f"{_normalize_directory(request.directory)}/{request.file_name}"
    target_path = _resolve_under_root(file_path)

    # Ensure directory exists
    os.makedirs(os.path.dirname(target_path), exist_ok=True)

    try:
        presentation.save(target_path)
    except Exception as exc:
        return error(f"Failed to create slides: {repr(exc)}")

    return CreateDeckResponse(
        success=True,
        file_name=request.file_name,
        file_path=file_path,
    )
